requirements.txt
Django>=5.0.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
django-environ>=0.10.0
django-redis>=5.4.0
django-rest-framework>=3.14.0
django-channels
channels-redis
python-dotenv>=1.0.0
whitenoise>=6.4.0
Pillow>=10.0.0
pdfkit>=1.0.0
reportlab>=4.0.0
chart.js>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
whitenoise>=6.4.0
psycopg2-binary>=2.9.9
